+++ chia/_tests/clvm/test_spend_sim.py
@@ -6,8 +6,9 @@ from chia_rs.sized_bytes import bytes32
 
 from chia._tests.util.spend_sim import sim_and_client
 from chia.types.blockchain_format.program import Program
-from chia.types.coin_spend import compute_additions, make_spend
+from chia.types.coin_spend import make_spend
 from chia.types.condition_opcodes import ConditionOpcode
+from chia.wallet.util.compute_additions import compute_additions
 
 
 @pytest.mark.anyio
